+++ CHANGES.md
@@ -9,6 +9,7 @@ Version 2.1.5 (Pending)
 - Update staff import & user creation to handle the new staff model
 - Add user creation for Ops volunteers during staff import
 - Change Staff positions from just role title to team + role title
+- Differentiate between guest and staff for badge text outline
 - 2024 Staff/Guest badges
 
 Version 2.1.4 (11/17/2023)